fix: made little fixes
